requirements.txt
fastapi
uvicorn
streamlit
requests
beautifulsoup4
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
tavily-python
langchain-groq